Move utility methods for DAP to the new file
