+++ test/support/test_case.rb
@@ -5,11 +5,13 @@ require 'tempfile'
 require 'securerandom'
 
 require_relative 'utils'
+require_relative 'dap_utils'
 require_relative 'assertions'
 
 module DEBUGGER__
   class TestCase < Test::Unit::TestCase
     include TestUtils
+    include DAP_TestUtils
     include AssertionHelpers
 
     def setup
